+++ CHANGELOG.md
@@ -1,9 +1,21 @@
+# Version 0.0.4
+
+## Fixes
+
+* DMLR-282: iOS not generating SvSR
+* DMLR-282: `Partner's name` field not editable
+* DMLR-282: missing auditor signature tab
+* DMLR-282: `Save and upload` button not calling survey uploading and not navigating to survey list
+
 # Version 0.0.3
-###Fixes
+
+## Fixes
 
 * fix(i18n): missed UI translations + survey filters fix [#49](https://git.daimler.com/GNSA/gnsa-mobile-audit-execution/pull/49)
 
 # Version 0.0.2
+
+## Fixes
 
 * Signature not able to draw consistent lines on real devices due to conflict with `ScrollView`
 * Enable hardware acceleration for `react-native-signature-canvas` on android